Bug #1704802 soc: sprd: debug: modify the format string
[root cause  ] the format string
[changes     ] sysdump
[side effects]no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:no
[board]:all
[test case]:yes
[reviewers   ]
[change_type ] feature_bugfix
[tag_product ] common

+++ drivers/soc/sprd/debug/sysdump/native_hang_monitor.c
@@ -38,11 +38,11 @@ static struct task_struct *hd_thread;
 
 static void log_to_hang_info(const char *fmt, ...)
 {
-	unsigned long len = 0;
+	int len = 0;
 	static int times;
 	va_list ap;
 
-	if ((hang_info_index + MAX_STRING_SIZE) >= (unsigned long)HANG_INFO_MAX) {
+	if ((hang_info_index + MAX_STRING_SIZE) >= (int)HANG_INFO_MAX) {
 		if (!times)
 			pr_err("HangInfo Buffer overflow len(0x%x), times:%d, HANG_INFO_MAX:0x%lx !!!!!!!\n",
 				hang_info_index, times, (long)HANG_INFO_MAX);
